+++ CHANGELOG.md
@@ -33,6 +33,7 @@ straightforward as possible.
 - Added list of comic per character
 - Added functionality "add favorites# in card.component
 - Added module comic's, in this module you can see the comic's of a character
+- Added pagination in the card.component
  
 ### Changed
 - Changed in the menu the item favorites, now it's acceded since home in the option "View all"